Additional complexity analysis added
Developed scripts to run library complexity analysis using preseq,
which is more accurate than the Lander-Waterman equation.  Rolled this
into the juicer pipeline as well as FastQC and FastQ Screen analysis.
Preseq analysis requires a lot of memory for large datasets, so this
may need to be tweaked in the future.